pyproject.toml
[tool.poetry]
name = "redditbot"
version = "0.7.5"
description = "Simple bot to get Reddit posts"
authors = ["Lucas <Polo>"]
license = "GPL V3"

[tool.poetry.dependencies]
python = "~3.12.0"
python-telegram-bot = "20.6"
click = "8.1.7"
dynaconf = "3.2.3"
httpx = "0.25.0"

[tool.poetry.group.dev.dependencies]
flake8 = "6.1.0"
flake8-quotes = "3.3.2"
pytest = "7.4.2"
pytest-cov = "4.1.0"
pytest-asyncio = "0.21.1"
pytest-httpx = "0.25.0"

[tool.pytest.ini_options]
asyncio_mode = 'auto'
addopts = "--cov-report term-missing --cov-report xml"

[tool.coverage.run]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = ["if __name__ == .__main__.:"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
